moar progress

+++ clean-compile-run.sh
@@ -21,7 +21,9 @@ cmake -S .. -DCMAKE_BUILD_TYPE=Debug -DCMAKE_EXPORT_COMPILE_COMMANDS=1
 make
 popd
 
-rm ./compile_commands.json
+if [ -L ./compile_commands.json ]; then
+  rm ./compile_commands.json
+fi
 ln -s ./build/compile_commands.json ./compile_commands.json
 
 ./build/vulking
